# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Waste Management API
  description: API for managing dropping points, prices, pickup orders, and user roles
  version: 1.0.0
  contact:
    name: API for Country Material Limited
    email: development@countrymaterial.com
servers:
  - url: http://54.209.99.13
    description: Development server
  - url: https://api.countrymaterial.com
    description: Production server
paths:
...
  /api/user-roles:
    post:
      summary: Assign a role to a user
      tags:
        - User Roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - role
              properties:
                user_id:
                  type: string
                  description: ID of the user
                  example: "user-123"
                role:
                  type: string
                  enum: [vendor, manager, admin]
                  description: Role to assign to the user
                  example: "vendor"
      responses:
        '201':
          description: User role created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User role created successfully"
        '400':
          description: Bad request - invalid role or missing fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid role. Must be: vendor, manager, or admin"
        '500':
          description: Internal server error
...
  /api/daily-price:
    post:
      summary: Create daily price for a dropping point
      tags:
        - Daily Prices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - dropping_point_id
                - category
                - price
                - created_by
              properties:
                dropping_point_id:
                  type: integer
                  description: ID of the dropping point
                  example: 1
                category:
                  type: string
                  description: Waste category
                  example: "heavy"
                price:
                  type: number
                  format: float
                  description: Price for the category
                  example: 150.50
                created_by:
                  type: string
                  description: ID of the user creating the price
                  example: "user-123"
      responses:
        '201':
          description: Daily price created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Daily price created successfully"
        '400':
          description: Bad request - missing required fields
        '500':
          description: Internal server error
...
  /api/pickup-order:
    post:
      summary: Create a new pickup order
      tags:
        - Pickup Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - vendor_id
                - dropping_point_id
                - category
                - price
                - phone_number
                - quantity
              properties:
                vendor_id:
                  type: string
                  description: ID of the vendor
                  example: "vendor-123"
                dropping_point_id:
                  type: integer
                  description: ID of the dropping point
                  example: 1
                category:
                  type: string
                  description: Waste category
                  example: "heavy"
                price:
                  type: number
                  format: float
                  description: Price per unit
                  example: 150.50
                phone_number:
                  type: string
                  description: Contact phone number
                  example: "+1234567890"
                quantity:
                  type: number
                  format: float
                  description: Quantity of waste
                  example: 10.5
                comment:
                  type: string
                  description: Additional comments
                  example: "Fragile items"
                image:
                  type: string
                  description: Base64 encoded image or image URL
                  example: "data:image/jpeg;base64,..."
      responses:
        '201':
          description: Pickup order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Pickup order created successfully"
        '400':
          description: Bad request - missing required fields
        '500':
          description: Internal server error
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []